+++ deps.ts
@@ -1,0 +1,33 @@
+import * as Glob from "https://deno.land/std@0.223.0/fs/expand_glob.ts";
+import * as SemVer from "https://deno.land/std@0.223.0/semver/mod.ts";
+
+import
+{
+    decodeBase64,
+    encodeBase64,
+} from "https://deno.land/std@0.223.0/encoding/base64.ts";
+
+import
+{
+    decodeBase32,
+    encodeBase32,
+} from "https://deno.land/std@0.224.0/encoding/base32.ts";
+
+import { Md5 } from "https://deno.land/std@0.71.0/hash/md5.ts";
+
+// @deno-types="npm:@types/node"
+import * as fs from "node:fs";
+import * as path from "node:path";
+import * as os from "node:os";
+
+export
+{
+    Glob, SemVer,
+    decodeBase64,
+    encodeBase64,
+    encodeBase32,
+    decodeBase32,
+    Md5,
+    fs, path, os,
+}
+