+++ source/compat.ts
@@ -1,21 +1,14 @@
 import
+  {
+    decodeBase32, decodeBase64, encodeBase32, encodeBase64,
+    Md5,
+    fs, os, path
+  } from '../deps.ts';
+export
 {
-  decodeBase64,
-  encodeBase64,
-} from "https://deno.land/std@0.223.0/encoding/base64.ts";
-import
-{
-  decodeBase32,
-  encodeBase32,
-} from "https://deno.land/std@0.224.0/encoding/base32.ts";
-import { Md5 } from "https://deno.land/std@0.71.0/hash/md5.ts";
-// @deno-types="npm:@types/node"
-import * as fs from "node:fs";
-import * as path from "node:path";
-import * as os from "node:os";
-
-export { decodeBase32, decodeBase64, encodeBase32, encodeBase64 };
-export { fs, os, path };
+  decodeBase32, decodeBase64, encodeBase32, encodeBase64,
+  fs, os, path
+};
 
 export class Md5Hasher
 {