+++ ChangeLog.md
@@ -1,3 +1,9 @@
+v0.1.6
+--------
+
+- Implement `Path.glob` (both static and non-static) (e7305392) based on [https://deno.land/std@0.223.0/fs/expand_glob.ts](https://deno.land/std@0.223.0/fs/expand_glob.ts).
+- Support `branch` parameter for `repoTarget` and `Repo` class methods.
+
 v0.1.5
 --------